Add new classes for OOP (Parser, Storage, TaskList, Ui) Add save method that will save at specified location
